Move scan_map_files into visualize.c, delete map_utils.h
